backend/requirements.txt
# --- Backend (FastAPI) ---
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
werkzeug

pandas
aiofiles
pypdf
huggingface-hub


llama-index-core
llama-index-readers-file

llama-index-llms-ollama

llama-index-embeddings-huggingface

torch
transformers
accelerate
bitsandbytes
protobuf
scipy